# openapi.yaml
openapi: 3.0.3
info:
  title: Manage User TODOs API
  version: 1.0.0
  description: |
    REST API for a multi-tenant TODO management platform. Users can register, authenticate,
    manage their profile, and perform CRUD operations on personal TODO items with tagging
    and reminder support.
servers:
  - url: https://api.todo.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
tags:
  - name: Auth
    description: Endpoints for user registration and authentication
  - name: Todos
    description: Manage personal TODO items
  - name: Users
    description: Manage the authenticated user's profile
  - name: UserDetails
    description: Capture custom user detail records tied to the authenticated user
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User created and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      tags: [Users]
      summary: Get profile of the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Users]
      summary: Update profile of the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user-details:
    get:
      tags: [UserDetails]
      summary: List all saved user detail entries for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Collection of saved entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [UserDetails]
      summary: Create a new user detail entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDetailInput'
      responses:
        '201':
          description: Entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /todos:
    get:
      tags: [Todos]
      summary: List TODOs for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/TagParam'
        - $ref: '#/components/parameters/CategoryParam'
      responses:
        '200':
          description: Paginated TODOs
          headers:
            X-Total-Count:
              description: Total number of TODOs for the query
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Todos]
      summary: Create a new TODO
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '201':
          description: TODO created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /todos/{todoId}:
    parameters:
      - name: todoId
        in: path
        required: true
        schema:
          type: string
        description: Unique identifier of the TODO
    get:
      tags: [Todos]
      summary: Retrieve a specific TODO
      security:
        - bearerAuth: []
      responses:
        '200':
          description: TODO details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Todos]
      summary: Update selected fields of a TODO
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Updated TODO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Todos]
      summary: Delete a TODO
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of TODOs per page
    StatusParam:
      name: status
      in: query
      schema:
        type: string
        enum: [pending, in_progress, done]
      description: Filter TODOs by status
    TagParam:
      name: tag
      in: query
      schema:
        type: string
      description: Filter TODOs by a given tag
    CategoryParam:
      name: category
      in: query
      schema:
        type: string
      description: Filter TODOs by category
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    UserRegistrationRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        fullName:
          type: string
          minLength: 2
        timezone:
          type: string
          description: IANA timezone identifier
    UserLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT access token
        refreshToken:
          type: string
          description: Refresh token to obtain new access tokens
        expiresIn:
          type: integer
          description: Access token lifetime in seconds
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      required: [id, email, fullName, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        avatarUrl:
          type: string
          format: uri
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserProfileRequest:
      type: object
      properties:
        fullName:
          type: string
        avatarUrl:
          type: string
          format: uri
        timezone:
          type: string
      minProperties: 1
    Todo:
      type: object
      required: [id, title, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
          nullable: true
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, done]
        dueDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        reminderAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TodoInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        category:
          type: string
          nullable: true
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        reminderAt:
          type: string
          format: date-time
    TodoUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          nullable: true
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, done]
    UserDetail:
      type: object
      required: [id, ownerId, fullName, email, createdAt, updatedAt]
      properties:
        id:
          type: string
        ownerId:
          type: string
          description: Identifier of the authenticated user that created the entry
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        department:
          type: string
        role:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserDetailInput:
      type: object
      required: [fullName, email]
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
          description: Optional contact number
        department:
          type: string
          description: Functional area or organizational unit
        role:
          type: string
          description: How the person engages with the product or institution
        notes:
          type: string
          maxLength: 500
        dueDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        reminderAt:
          type: string
          format: date-time
      minProperties: 1
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
        details:
          type: array
          items:
            type: string